Added closed loop

+++ src/main/java/frc/robot/Constants.java
@@ -49,8 +49,8 @@ public final class Constants {
             public static final double kD = 0.0;
             public static final double kIz = 0.0;
             public static final double kFF = 0.0;
-            public static final int kMaxOutput = 1; 
-            public static final int kMinOutput = -1;
+            public static final double kMaxOutput = 1; 
+            public static final double kMinOutput = -1;
             public static final int maxRPM = 5700;
             
         }
